docs/readthedocs.requirements.txt
mkdocs>=1.1.2
mkdocs-material>=6.1.0
mkdocstrings>=0.13.6
jinja2>=3.1.0
setuptools>=65.0.0
numpy>=1.20.0
pandas>=1.0.0
scipy>=1.5.0
scikit-learn>=0.24.0
PyWavelets>=1.1.0
pytest>=4.0.0
pytest-cov>=2.10.0
joblib>=0.16.0
jsonschema>=3.2.0
kennard-stone>=0.5.0
twinning>=0.0.5
optuna>=2.0.0
tensorflow>=2.0.0
torch>=1.4.0
nirs4all @ git+https://github.com/GBeurier/nirs4all
sphinx_rtd_theme
myst-parser
sphinx-copybutton
